Move videos formatting to useFetchVideosInfinite
